cloudtests: move cloud resource to separate files
