+++ misc/python/materialize/cloudtest/k8s/debezium.py
@@ -22,7 +22,8 @@ from kubernetes.client import (
     V1ServiceSpec,
 )
 
-from materialize.cloudtest.k8s import K8sDeployment, K8sService
+from materialize.cloudtest.k8s.api.k8s_deployment import K8sDeployment
+from materialize.cloudtest.k8s.api.k8s_service import K8sService
 
 
 class DebeziumDeployment(K8sDeployment):
