+++ misc/python/materialize/cloudtest/k8s/vpc_endpoints_cluster_role.py
@@ -9,7 +9,7 @@
 
 from kubernetes.client import V1ClusterRole, V1ObjectMeta, V1PolicyRule
 
-from materialize.cloudtest.k8s import K8sClusterRole
+from materialize.cloudtest.k8s.api.k8s_cluster_role import K8sClusterRole
 
 
 class VpcEndpointsClusterRole(K8sClusterRole):
